pyproject.toml
[build-system]
requires = ["setuptools>=66,<68", "wheel"] # , "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "PowerGenome"
version = "0.6.1a2"
authors = [{ name = "Greg Schivley", email = "greg.schivley@princeton.edu" }]
maintainers = [
    { name = "Greg Schivley", email = "greg.schivley@princeton.edu" },
]
description = "Create power system inputs for capacity expansion models"
readme = "README.md"
requires-python = ">=3.8"
keywords = ["power system data", "capacity expansion", "two"]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "catalystcoop.pudl>=0.6.0, <=2022.11.30",
    "beautifulsoup4 >= 4.8.13",
    "statsmodels>=0.12.2",
    "python-dotenv",
    "flatten-dict",
    "ruamel.yaml",
    "pyyaml",
    "frozendict",
    "openpyxl>=3.0",
    "geopandas>=0.11",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["powergenome*", "data*"]
exclude = ["tests*", "docs*", "data/eia*"]

[tool.setuptools.package-data]
"*" = ["*.csv", "*.geojson"]

[project.optional-dependencies]
dev = [
    "black[jupyter] == 23.3.0",
    "pre-commit",
    "pytest",
    "isort",
    "build",
    "twine",
]

[project.scripts]
run_powergenome_multiple = "powergenome.run_powergenome_multiple_outputs_cli:main"

[project.urls]
"Source" = "https://github.com/PowerGenome/PowerGenome"
